test(set/set_tests): move and rewrite to work in tests
<body>

<footer>
